130 tons of new stuff. Major stuff: Scripting Bug hunting

+++ src/main/java/dev/heliosclient/managers/CapeManager.java
@@ -2,17 +2,15 @@ package dev.heliosclient.managers;
 
 import com.google.gson.*;
 import dev.heliosclient.HeliosClient;
-import dev.heliosclient.event.SubscribeEvent;
-import dev.heliosclient.event.events.TickEvent;
-import dev.heliosclient.event.listener.Listener;
+import dev.heliosclient.module.modules.misc.CapeModule;
 import dev.heliosclient.system.HeliosExecutor;
 import dev.heliosclient.util.ColorUtils;
 import dev.heliosclient.util.animation.AnimationUtils;
-import dev.heliosclient.util.cape.CapeSynchronizer;
 import dev.heliosclient.util.cape.ProfileUtils;
 import dev.heliosclient.util.fontutils.FontLoader;
-import dev.heliosclient.util.render.GifTexture;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.render.WorldRenderer;
+import net.minecraft.client.render.entity.EntityRenderDispatcher;
 import net.minecraft.client.texture.NativeImage;
 import net.minecraft.client.texture.NativeImageBackedTexture;
 import net.minecraft.client.texture.TextureManager;
@@ -29,33 +27,27 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.*;
+import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
-public class CapeManager implements Listener {
+public class CapeManager {
     public static CapeManager INSTANCE = new CapeManager();
-
     private static final File CAPE_DIRECTORY = new File(HeliosClient.MC.runDirectory, "heliosclient/capes");
     private static final String DEFAULT_CAPE = "helioscape.png";
-    public static final Identifier CAPE_TEXTURE = new Identifier("heliosclient", "capes/" + DEFAULT_CAPE);
     public static final Identifier DEFAULT_CAPE_TEXTURE = new Identifier("heliosclient", "capes/" + DEFAULT_CAPE);
-    private static final Map<UUID, Identifier> CAPES = new HashMap<>();
-    private static final Map<UUID, Identifier> ELYTRAS = new HashMap<>();
     public static String[] capes = new String[]{};
     public static Identifier cape;
+
+    private static final Set<String> registeredTextures = new HashSet<>();
+    private static final Map<UUID, Identifier> CAPES = new HashMap<>();
+    private static final Map<UUID, Identifier> ELYTRAS = new HashMap<>();
+
     public static List<Identifier> capeIdentifiers = new ArrayList<>();
     public static List<Identifier> elytraIdentifiers = new ArrayList<>();
 
-    @SubscribeEvent
-    public void onTick(TickEvent e) {
-        if (HeliosClient.MC.getWindow() != null) {
-            CapeManager.capes = CapeManager.loadCapes();
-            CapeSynchronizer.registerCapeSyncPacket();
-            EventManager.unregister(this);
-        }
-    }
     /**
-     * Works similar to FontLoader#loadFonts
+     * Works similar to {@link FontLoader#loadFonts()}
      */
     public static String[] loadCapes() {
         Future<String[]> future = HeliosExecutor.submit(() -> {
@@ -67,27 +59,29 @@ public class CapeManager implements Listener {
         }
 
         File defaultCapeFile = new File(CAPE_DIRECTORY, DEFAULT_CAPE);
-        try (InputStream inputStream = FontLoader.class.getResourceAsStream("/assets/heliosclient/capes/helioscape.png")) {
-            if (inputStream == null) {
-                HeliosClient.LOGGER.error("Failed to load open inputStream to default cape resource");
-            } else {
-                HeliosClient.LOGGER.info("Copying default cape in directory");
-                Files.copy(inputStream, defaultCapeFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-                HeliosClient.LOGGER.info("Copying completed");
+
+        // Do not copy if the default cape file already exists
+        if(!defaultCapeFile.exists()) {
+            try (InputStream inputStream = FontLoader.class.getResourceAsStream("/assets/heliosclient/capes/helioscape.png")) {
+                if (inputStream == null) {
+                    HeliosClient.LOGGER.error("Failed to load open inputStream to default cape resource");
+                } else {
+                    HeliosClient.LOGGER.info("Copying default cape in directory");
+                    Files.copy(inputStream, defaultCapeFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                    HeliosClient.LOGGER.info("Copying completed");
+                }
+            } catch (IOException e) {
+                HeliosClient.LOGGER.error("An error has occured while reading default resource asset cape", e);
             }
-        } catch (IOException e) {
-            HeliosClient.LOGGER.error("An error has occured while reading default resource asset cape", e);
         }
 
-        File[] capeFiles = CAPE_DIRECTORY.listFiles((dir, name) -> name.toLowerCase().endsWith(".png") || name.toLowerCase().endsWith(".gif"));
+        // Get the cape files from `heliosclient/capes` directory in an array
+        File[] capeFiles = CAPE_DIRECTORY.listFiles((dir, name) -> name.toLowerCase().endsWith(".png"));
 
         if (capeFiles == null) {
             HeliosClient.LOGGER.info("No cape files found");
             return new String[0];
         }
-
-        capeIdentifiers.clear();
-        elytraIdentifiers.clear();
 
         List<String> capeNames = new ArrayList<>();
 
@@ -97,44 +91,45 @@ public class CapeManager implements Listener {
                 String capeName = fileName.substring(0, fileName.lastIndexOf('.'));
                 capeNames.add(capeName);
 
-                load(inputStream, fileName, file);
+                loadCapeTexture(inputStream, fileName, file);
             } catch (IOException e) {
                 HeliosClient.LOGGER.error("An error has occured while reading cape file: " + ColorUtils.darkGreen + file.getName(), e);
             }
         }
-        String[] capeNamesArray = capeNames.toArray(new String[0]);
-        capes = capeNamesArray;
-        return capeNamesArray;
+        capes = capeNames.toArray(new String[0]);
+        return capes;
     });
         try {
-            return future.get();
+            String[] result = future.get();
+            CapeModule.get().capes.options = List.of(result);
+            return result;
         } catch (InterruptedException | ExecutionException e) {
             HeliosClient.LOGGER.error("An error occurred while loading capes", e);
             return new String[0];
         }
     }
 
-    public static void load(InputStream inputStream, String fileName, File file) throws IOException {
+    public static void loadCapeTexture(InputStream inputStream, String fileName, File file) throws IOException {
+        if (registeredTextures.contains(fileName)) {
+            return; // Skip if texture is already registered
+        }
+
         NativeImage image = NativeImage.read(inputStream);
 
         HeliosClient.MC.execute(() -> {
-            try {
-                TextureManager textureManager = HeliosClient.MC.getTextureManager();
-                Identifier capeIdentifier;
-                Identifier elytraIdentifier;
-                if (fileName.endsWith(".gif")) {
-                    capeIdentifier = textureManager.registerDynamicTexture("cape_" + fileName.toLowerCase(Locale.ROOT), new GifTexture(file));
-                    elytraIdentifier = capeIdentifier;
-                } else {
-                    capeIdentifier = textureManager.registerDynamicTexture("cape_" + fileName.toLowerCase(Locale.ROOT), new NativeImageBackedTexture(parseCape(image)));
-                    elytraIdentifier = textureManager.registerDynamicTexture("elytra_" + fileName.toLowerCase(Locale.ROOT), new NativeImageBackedTexture(image));
-                }
-                elytraIdentifiers.add(elytraIdentifier);
-                capeIdentifiers.add(capeIdentifier);
-                HeliosClient.LOGGER.info("Loaded cape: " + fileName);
-            } catch (IOException e) {
-                HeliosClient.LOGGER.error("An error has occured while parsing and registering cape file: " + ColorUtils.darkGreen + file.getName(), e);
-            }
+            TextureManager textureManager = HeliosClient.MC.getTextureManager();
+            Identifier capeIdentifier;
+            Identifier elytraIdentifier;
+            capeIdentifier = textureManager.registerDynamicTexture("cape_" + fileName.toLowerCase(Locale.ROOT), new NativeImageBackedTexture(parseCape(image)));
+            elytraIdentifier = textureManager.registerDynamicTexture("elytra_" + fileName.toLowerCase(Locale.ROOT), new NativeImageBackedTexture(image));
+
+
+            capeIdentifiers.add(capeIdentifier);
+            elytraIdentifiers.add(elytraIdentifier);
+
+            registeredTextures.add(fileName);
+
+            HeliosClient.LOGGER.info("Loaded cape: " + fileName);
         });
     }
 
@@ -164,53 +159,44 @@ public class CapeManager implements Listener {
         return imgNew;
     }
 
-    public static void loadCape(PlayerEntity player, Identifier capeTexture) {
-        if (HeliosClient.MC.getResourceManager() != null) {
-            try (InputStream stream = MinecraftClient.getInstance().getResourceManager().getResource(capeTexture).orElseThrow().getInputStream()) {
-                NativeImage image = NativeImage.read(stream);
-                NativeImageBackedTexture texture = new NativeImageBackedTexture(image);
-                MinecraftClient.getInstance().getTextureManager().registerTexture(capeTexture, texture);
-                CapeManager.setCapeAndElytra(player, capeTexture, capeTexture);
-            } catch (IOException e) {
-                e.printStackTrace();
-                // If the cape texture fails to load, set the player's cape to the default cape
-                CapeManager.setCapeAndElytra(player, DEFAULT_CAPE_TEXTURE, DEFAULT_CAPE_TEXTURE);
+    public static void getCapes(CapeType type, String profileName, String UUID) throws Exception {
+        Future<?> future = HeliosExecutor.submit((Callable<Void>) () -> {
+            switch (type) {
+                case OPTIFINE:
+                    if (profileName == null || profileName.isEmpty()) {
+                        throw new IllegalArgumentException("Profile name is required for OPTIFINE capes.");
+                    }
+                    // Get the optifine cape
+                    INSTANCE.getOptifineCape(profileName);
+                    break;
+                case CRAFATAR:
+                case MINECRAFTCAPES:
+                    if (UUID == null || UUID.isEmpty()) {
+                        throw new IllegalArgumentException("Complete and a valid UUID is required.");
+                    }
+                    if (!ProfileUtils.isValidUUID(UUID)) {
+                        throw new IllegalArgumentException("Invalid UUID");
+                    }
+                    if (type == CapeType.CRAFATAR) {
+                        // Get the craftar cape
+                        INSTANCE.getCrafatarCape(UUID);
+                    } else {
+                        // Get the minecraft cape
+                        INSTANCE.getMinecraftCapesCape(UUID);
+                    }
+                    break;
+                case NONE:
+                    return null;
+                default:
+                    throw new IllegalArgumentException("Invalid cape type: " + type);
             }
-        }
+            loadCapes();
+            return null;
+        });
+        future.get();
     }
 
-    // No executors here because fuck I like lag sometimes
-    public static void getCapes(CapeType type, String profileName, String UUID) throws IOException {
-        switch (type) {
-            case OPTIFINE:
-                if (profileName == null || profileName.isEmpty()) {
-                    throw new IllegalArgumentException("Profile name is required for OPTIFINE capes.");
-                }
-                INSTANCE.getOptifineCape(profileName);
-                break;
-            case CRAFATAR:
-            case MINECRAFTCAPES:
-                if (UUID == null || UUID.isEmpty()) {
-                    throw new IllegalArgumentException("Valid UUID is required for CRAFATAR and MINECRAFTCAPES capes.");
-                }
-                if (!ProfileUtils.isValidUUID(UUID)) {
-                    throw new IllegalArgumentException("Invalid UUID");
-                }
-                if (type == CapeType.CRAFATAR) {
-                    INSTANCE.getCrafatarCape(UUID);
-                } else {
-                    INSTANCE.getMinecraftCapesCape(UUID);
-                }
-                break;
-            case NONE:
-                return;
-            default:
-                throw new IllegalArgumentException("Invalid cape type: " + type);
-        }
-        loadCapes();
-    }
-
-    private static void saveCapeFromBase64(String UUID, String url) throws IOException {
+    private static void saveCapeFromBase64(String UUID, String url) throws Exception {
         HttpURLConnection connection = INSTANCE.getConnection(url);
         connection.connect();
 
@@ -219,18 +205,26 @@ public class CapeManager implements Listener {
             JsonObject result = new Gson().fromJson(reader, JsonObject.class);
 
             if (result == null) {
-                throw new JsonParseException("Json returned is null");
+                throw new JsonParseException("Json result returned is null");
             }
             JsonElement jsonElement = result.getAsJsonObject("textures").get("cape");
-            if (jsonElement.isJsonNull()) {
-                HeliosClient.LOGGER.error("UUID does not contain any capes: " + UUID + " If you are sure that this UUID should contain a cape then try other service (i.e Craftar or Minecraft capes");
-                AnimationUtils.addErrorToast("UUID does not contain any capes Check Logs for details", true, 1500);
+            boolean animated = result.getAsJsonObject("animatedCape").getAsBoolean();
+
+            if(animated){
+                HeliosClient.LOGGER.error("Animated capes are not supported");
+                AnimationUtils.addErrorToast("Animated capes are not supported", false, 1500);
                 connection.disconnect();
                 return;
             }
+
+            if (jsonElement.isJsonNull()) {
+                HeliosClient.LOGGER.error("UUID does not contain any capes: " + UUID + " If you are sure that this UUID should contain a cape then try other service (i.e Craftar or Minecraft capes");
+                AnimationUtils.addErrorToast("UUID does not contain any capes, Check Logs for details", true, 1500);
+                connection.disconnect();
+                throw new NullPointerException("UUID does not contain any capes");
+            }
+
             String capeTexture = jsonElement.getAsString();
-
-
             byte[] imageBytes = Base64.getDecoder().decode(capeTexture);
             BufferedImage image = ImageIO.read(new ByteArrayInputStream(imageBytes));
 
@@ -244,22 +238,22 @@ public class CapeManager implements Listener {
         }
     }
 
-    private void getOptifineCape(String profileName) throws IOException {
+    private void getOptifineCape(String profileName) throws Exception {
         String url = "http://s.optifine.net/capes/" + profileName + ".png";
         saveCapeFromUrl(profileName, null, url);
     }
 
-    private void getCrafatarCape(String UUID) throws IOException {
+    private void getCrafatarCape(String UUID) throws Exception {
         String url = "https://crafatar.com/capes/" + UUID;
         saveCapeFromUrl(null, UUID, url);
     }
 
-    private void getMinecraftCapesCape(String UUID) throws IOException {
+    private void getMinecraftCapesCape(String UUID) throws Exception {
         String url = "https://api.minecraftcapes.net/profile/" + UUID.replace("-", "");
         saveCapeFromBase64(UUID, url);
     }
 
-    public HttpURLConnection getConnection(String url) throws IOException {
+    public HttpURLConnection getConnection(String url) throws Exception {
         HttpURLConnection connection = (HttpURLConnection) new URL(url).openConnection(MinecraftClient.getInstance().getNetworkProxy());
         connection.setRequestProperty("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36");
         connection.setDoInput(true);
@@ -267,7 +261,7 @@ public class CapeManager implements Listener {
         return connection;
     }
 
-    private void saveCapeFromUrl(String profileName, String UUID, String url) throws IOException {
+    private void saveCapeFromUrl(String profileName, String UUID, String url) throws Exception {
         HttpURLConnection connection = INSTANCE.getConnection(url);
         connection.connect();
 
@@ -284,7 +278,6 @@ public class CapeManager implements Listener {
             throw new HttpResponseException(connection.getResponseCode(),connection.getResponseMessage());
         }
     }
-
     public static boolean shouldPlayerHaveCape(PlayerEntity player) {
         return CAPES.containsKey(player.getUuid()) || ELYTRAS.containsKey(player.getUuid());
     }
@@ -303,9 +296,9 @@ public class CapeManager implements Listener {
     }
 
     public enum CapeType {
+        NONE,
         OPTIFINE,
         CRAFATAR,
-        MINECRAFTCAPES,
-        NONE
+        MINECRAFTCAPES
     }
 }
